+++ index.d.ts
@@ -21,6 +21,7 @@ export type Result
 export function bind(target: DocumentFragment | HTMLElement | ShadowRoot, settings: ParserSettings): {
   parse: (source: string) => Generator<Result, undefined, undefined>;
   nearest: (position: number) => HTMLElement | undefined;
+  position: (block: HTMLElement) => number;
 };
 export function render(target: HTMLElement, options?: RenderingOptions): void;
 export function quote(address: string, range: Range): string;
@@ -28,7 +29,7 @@ export function toc(source: DocumentFragment | HTMLElement | ShadowRoot): HTMLUL
 export function info(source: DocumentFragment | HTMLElement | ShadowRoot): Info;
 export function context(base: DocumentFragment | HTMLElement | ShadowRoot, bound?: string): (el: Element) => boolean;
 export function normalize(source: string): string;
-export function sync(editor: HTMLElement, viewer: HTMLElement, footnote: HTMLOListElement[]): () => void;
+export function sync(editor: HTMLElement, viewer: HTMLElement, footnote: readonly HTMLOListElement[]): () => void;
 
 export type ParserOptions = Partial<ParserSettings>;
 export interface ParserSettings {