+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.191.0
 
 - Refine bind function.
+- Add sync function.
 - Drop media support of Gist for keeping word limits of timelines and threads.
 
 ## 0.190.8